+++ src/Haskellorls/SymbolicLink.hs
@@ -12,12 +12,13 @@ import qualified Haskellorls.Name.Decorator as Name
 import qualified Haskellorls.NodeInfo as Node
 import qualified Haskellorls.Option as Option
 import qualified Haskellorls.Quote.Utils as Quote
+import qualified Haskellorls.Utils as Utils
 import qualified Haskellorls.WrappedText as WT
 
 linkName :: Option.Option -> Node.NodeInfo -> [WT.WrappedText]
 linkName opt node = case node of
   Node.FileInfo {} -> []
-  _ -> prefix' <> Quote.quote style (WT.toWrappedText link)
+  _ -> prefix' <> Quote.quote style (WT.toWrappedText $ Utils.escapeFormatter opt link)
   where
     style = Quote.quoteStyleForLink opt
     link = Name.nodeName $ Node.toFileInfo node
@@ -29,7 +30,7 @@ coloredLinkName opt config node = case node of
   _ -> prefix' <> Quote.quote style link
   where
     style = Quote.quoteStyleForLink opt
-    link = Name.colorizedNodeName config $ Node.toFileInfo node
+    link = Name.colorizedNodeName opt config $ Node.toFileInfo node
     prefix' = WT.toWrappedTextSingleton prefix
 
 prefix :: T.Text